simplified reshaping in amplitude_scattering_matrix()
We now assume thetas has shape (..., num_thetas) and that we broadcast over phi
if specified.  This change will affect downstream code that calculates the
amplitude scattering matrix for all pairs (theta_ij, phi_ij) but is needed to
allow theta and phi to be specified as a function of wavelength.